% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Combine_Feature.R
\name{Combine_Feature}
\alias{Combine_Feature}
\title{Combine Image features}
\usage{
Combine_Feature(cwd, fpath_txt)
}
\arguments{
\item{cwd}{Filepath of project directory}

\item{fpath_txt}{Filepath of .txt containing image information}
}
\value{
Save combined features. None is returned
}
\description{
Combine Image features
}
