% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Combine_Feature_v2.R
\name{Combine_Feature_v2}
\alias{Combine_Feature_v2}
\title{Combine image features according to each model}
\usage{
Combine_Feature_v2(cwd)
}
\arguments{
\item{cwd}{Filepath of project directory}
}
\value{
Save combined features. None is returned
}
\description{
Combine image features according to each model
}
\examples{
Combine_Feature_v2("/Users/benchiang/Documents/ComputationalSocialScience/R_Package_Material")
}
