% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicate_Image_Kmeans.R
\name{Duplicate_Image}
\alias{Duplicate_Image}
\title{Duplicate images with specific label}
\usage{
Duplicate_Image(label_20, model, dir, cwd)
}
\arguments{
\item{label_20}{Dataframe containing images with specific label}

\item{model}{Model name}

\item{dir}{Directory name(e.g. 6_Means)}

\item{cwd}{Filepath of project directory}
}
\description{
Duplicate images with specific label
}
