% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicate_Image.R
\name{Duplicate_Image_Kmeans}
\alias{Duplicate_Image_Kmeans}
\title{Duplicate images for all images clustered by K-means methods}
\usage{
Duplicate_Image_Kmeans(cwd)
}
\arguments{
\item{cwd}{File path of project directory}
}
\value{
Call Duplicate_Image() to save images with specific label.
}
\description{
Duplicate images for all images clustered by K-means methods
}
