% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract_Feature.R
\name{Extract_Feature}
\alias{Extract_Feature}
\title{Extract image features}
\usage{
Extract_Feature(model, base_model_name, cwd, fpath_txt, delim, header)
}
\arguments{
\item{model}{character Loaded learning model}

\item{base_model_name}{Name of base model}

\item{cwd}{Filepath of project directory}

\item{fpath_txt}{Filepath of .txt containing image information}

\item{delim}{Delimiter of .txt file}

\item{header}{Whether .txt file has header}
}
\value{
extracted features are saved in .dat files. None is returned
}
\description{
Extract image features
}
\examples{
Extract_Feature(model16, "VGG16_fc1", "/Users/benchiang/Documents/ComputationalSocialScience/R_Package_Material","/Users/benchiang/Documents/ComputationalSocialScience/R_Package_Material/img filename study2A.txt",delim='\t', header=TRUE)
}
