% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_txt.R
\name{Generate_txt}
\alias{Generate_txt}
\title{Generates .txt file containing image file paths from their folder(s).}
\usage{
Generate_txt(cwd)
}
\arguments{
\item{cwd}{Filepath of the project directory containing img file}
}
\description{
Generates .txt file containing image file paths from their folder(s).
}
\examples{
Generate_txt("/Users/benchiang/Documents/ComputationalSocialScience/R_Package_Material")
}
