% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_Image.R
\name{Image_Processing}
\alias{Image_Processing}
\title{Load, resize images to desirable dimensions, and save them as a specific file type(e.g. jpg)}
\usage{
Image_Processing(img_filepath, img_type)
}
\arguments{
\item{img_dimension(character)}{The dimension(widthxheight) of a resized image. The default is 0 -- unchanged.
Note: The image size may be changed permanently if the image is already in the desirable file type.}

\item{img_filepath(character)}{The file path of the project directory}

\item{img_type(character)}{The desirable image file type(e.g. jpg, jpeg, png, and gif)}
}
\value{
Images will be saved. None will be returned
}
\description{
Load, resize images to desirable dimensions, and save them as a specific file type(e.g. jpg)
}
\examples{
process_img("/Users/benchiang/Documents/ComputationalSocialScience/R_Package_Material", "jpg")
process_img("/Users/benchiang/Documents/ComputationalSocialScience/R_Package_Material", "jpeg" )
}
