% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Install_Py_Packages_v2.R
\name{Install_Conda}
\alias{Install_Conda}
\title{Install Python packages via Conda}
\usage{
Install_Conda(envname_conda, pac_c)
}
\arguments{
\item{envname_conda}{The name of, or path to, a conda environment}

\item{pac_c}{Character vector containing list of packages to be installed}
}
\description{
The parameter is based on reticulate package.(https://rstudio.github.io/reticulate/reference/conda-tools.html)
}
\examples{
Install_Conda("my_conda_env",c("joblib","numpy","pandas","sklearn","pillow","keras","keras_applications","keras-vggface","tensorflow"))
}
