% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Install_Py_Packages_v2.R
\name{Install_Default}
\alias{Install_Default}
\title{Install Python packages by default(using miniconda)}
\usage{
Install_Default(pac_c)
}
\arguments{
\item{pac_c}{Character vector containing list of packages to be installed}
}
\description{
Install Python packages by default(using miniconda)
}
\examples{
Install_Default()
}
