% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Install_Py_Packages_v2.R
\name{Install_Py_Packages_v2}
\alias{Install_Py_Packages_v2}
\title{Check whether the Python packages are installed. If not, installation will be carried out.}
\usage{
Install_Py_Packages_v2(
  method = "default",
  envname = "my_env",
  python = NULL,
  envname_conda = "my_conda_env"
)
}
\arguments{
\item{method}{Installation method}

\item{envname}{See Install_Virtualenv(envname="my_env", python=NULL) below}

\item{python}{See Install_Virtualenv(envname="my_env", python=NULL) below}

\item{envname_conda}{See Install_Conda(envname_conda,pac_c) below}
}
\description{
Check whether the Python packages are installed. If not, installation will be carried out.
}
\examples{
Install_Py_Packages_v2("default")
Install_Py_Packages_v2("virtualenv",envname="my_env")
Install_Py_Packages_v2("conda",envname="~/Documents/conda_test")
}
