% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Install_Py_Packages_v2.R
\name{Install_Virtualenv}
\alias{Install_Virtualenv}
\title{Install Python packages via virtualenv}
\usage{
Install_Virtualenv(pac_c, envname = "my_env", python = "mypython_path")
}
\arguments{
\item{pac_c}{Character vector containing list of packages to be installed}

\item{envname}{The name of, or path to, a Python virtual environment. If this name contains any slashes, the name will be interpreted as a path; if the name does not contain slashes, it will be treated as a virtual environment within virtualenv_root().
When NULL, the virtual environment as specified by the RETICULATE_PYTHON_ENV environment variable will be used instead. To refer to a virtual environment in the current working directory, you can prefix the path with ./<name>.}

\item{python}{The path to a Python interpreter, to be used with the created virtual environment. When NULL, the Python interpreter associated with the current session will be used. ()}
}
\description{
The parameters are based on reticulate packages. ((https://rstudio.github.io/reticulate/reference/virtualenv-tools.html))
}
\examples{
Install_Virtualenv(c("joblib","numpy","pandas","sklearn","pillow","keras","keras_applications","keras-vggface","tensorflow"),envname="my_env", python=NULL)
}
