% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_Means_Clustering.R
\name{K_Means_Clustering}
\alias{K_Means_Clustering}
\title{Performs K-Means clustering and save assigned label for each image as csv.}
\usage{
K_Means_Clustering(cwd, num_cluster, fpath_txt)
}
\arguments{
\item{cwd}{File path of project directory}

\item{num_cluster}{Number of clusters}

\item{fpath_txt}{File path of .txt containing image information}
}
\value{
Save assigned label for each image as csv.
}
\description{
Performs K-Means clustering and save assigned label for each image as csv.
}
