% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_Means_Clustering_v2.R
\name{K_Means_Clustering_v2}
\alias{K_Means_Clustering_v2}
\title{Performs K-Means clustering and save assigned label for each image as csv.}
\usage{
K_Means_Clustering_v2(cwd, num_cluster)
}
\arguments{
\item{cwd}{File path of project directory}

\item{num_cluster}{Number of clusters}
}
\value{
Save assigned label for each image as csv.
}
\description{
Performs K-Means clustering and save assigned label for each image as csv.
}
