% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load_Place365.R
\name{Load_Place365}
\alias{Load_Place365}
\title{Load VGG16_Place_365 model and print out model summary(e.g. output shapes and Param #)}
\usage{
Load_Place365()
}
\value{
a VGG16_Place_365 model
}
\description{
Load VGG16_Place_365 model and print out model summary(e.g. output shapes and Param #)
}
\examples{
Load_Place365()
}
