% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load_VGG16.R
\name{Load_VGG16}
\alias{Load_VGG16}
\title{Load VGG16 model and print out model summary(e.g. output shapes and Param #)}
\usage{
Load_VGG16()
}
\value{
a VGG16 model
}
\description{
Load VGG16 model and print out model summary(e.g. output shapes and Param #)
}
\examples{
Load_VGG16()
}
