% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load_VGGFace.R
\name{Load_VGGFace}
\alias{Load_VGGFace}
\title{Load VGGFace model and print out model summary(e.g. output shapes and Param #)}
\usage{
Load_VGGFace()
}
\value{
a VGGFace model
}
\description{
Load VGGFace model and print out model summary(e.g. output shapes and Param #)
}
\examples{
Load_VGGFace()
}
