% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{PCA}
\alias{PCA}
\title{Peform PCA}
\usage{
PCA(cwd)
}
\arguments{
\item{cwd}{Filepath of the project directory}
}
\description{
Peform PCA
}
\examples{
PCA("/Users/benchiang/Documents/ComputationalSocialScience/R_Package_Material")
}
