% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_Image.R
\name{Process_Image}
\alias{Process_Image}
\title{Save images in a specific file type(e.g. jpg)}
\usage{
Process_Image(img_filepath, img_type)
}
\arguments{
\item{img_filepath(character)}{File path of the project directory}

\item{img_type(character)}{The desirable image file type(e.g. jpg, jpeg, png, and gif)}
}
\description{
Save images in a specific file type(e.g. jpg)
}
\examples{
Process_Image("/Users/benchiang/Documents/ComputationalSocialScience/R_Package_Material", "jpg")
Process_Image("/Users/benchiang/Documents/ComputationalSocialScience/R_Package_Material", "jpeg" )
}
