% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Save_Img.R
\name{Save_Image}
\alias{Save_Image}
\title{Save images with same label. Called within Save_Img}
\usage{
Save_Image(cwd, jpg_list_df, k_val, model, w_sq, w_clusterid, h_imgid)
}
\arguments{
\item{cwd}{Filepath of project directory}

\item{jpg_list_df}{Information(filepath and label) for images}

\item{k_val}{K in k-means}

\item{model}{Model name}

\item{w_sq}{Width/height of image (300)}

\item{w_clusterid}{Width of text for cluster id (60)}

\item{h_imgid}{Height of gap (60)}
}
\description{
Save images with same label. Called within Save_Img
}
