% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subset_txt.R
\name{Subset_txt}
\alias{Subset_txt}
\title{Subset the .txt generated from Generate_txt function}
\usage{
Subset_txt(cwd, num_obs)
}
\arguments{
\item{cwd}{Filepath of the project directory containing img file}

\item{num_obs}{Number of observations used for testing. Note that}
}
\description{
Subset the .txt generated from Generate_txt function
}
\examples{
Subset_txt("/Users/benchiang/Documents/ComputationalSocialScience/R_Package_Material", 100)
}
