% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicate_Image_Kmeans.R
\name{Duplicate_Image_Kmeans}
\alias{Duplicate_Image_Kmeans}
\title{Duplicate images for all images clustered by K-means methods.
Within this function, we call Call Duplicate_Image() to save images with specific label.}
\usage{
Duplicate_Image_Kmeans(cwd)
}
\arguments{
\item{cwd}{File path of project directory}
}
\description{
Duplicate images for all images clustered by K-means methods.
Within this function, we call Call Duplicate_Image() to save images with specific label.
}
\examples{
Duplicate_Image_Kmeans("/Users/benchiang/Documents/ComputationalSocialScience/R_Package_Material")
}
