% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract_Feature_v2.R
\name{Extract_Feature_v2}
\alias{Extract_Feature_v2}
\title{Extract image features and save them as .dat files}
\usage{
Extract_Feature_v2(model, base_model_name, cwd, img_dimension)
}
\arguments{
\item{model}{character Loaded learning model}

\item{base_model_name}{Name of base model}

\item{cwd}{Filepath of project directory}

\item{img_dimension}{Column vector specifying width and height of resized images, e.g. c(224,224)}
}
\description{
Extract image features and save them as .dat files
}
\examples{
Extract_Feature(model16, "VGG16_fc1", "/Users/benchiang/Documents/ComputationalSocialScience/R_Package_Material", c(224,224))
}
